[write-fonts] Split up the splitting module
This is going to start getting unwieldly with the addition of Mark2Base
